# docs/openapi.yaml
openapi: 3.0.3
info:
  title: IgwebuIke Alumni Network API
  version: 2.3.0
  description: Secure alumni networking API. Authentication uses an httpOnly cookie named `iw_token`.
servers:
- url: https://igwebuike-backend-1.onrender.com
- url: http://localhost:5000
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: iw_token
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    RegisterRequest:
      type: object
      required:
      - firstName
      - lastName
      - email
      - password
      - gradYear
      - department
      - consent
      properties:
        firstName:
          type: string
          maxLength: 50
        lastName:
          type: string
          maxLength: 50
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 10
          maxLength: 72
          description: Max 72 BYTES (bcrypt limit)
        gradYear:
          type: integer
        department:
          type: string
        phone:
          type: string
          maxLength: 20
        location:
          type: string
        occupation:
          type: string
        consent:
          type: boolean
          description: Must be true
    LoginRequest:
      type: object
      required:
      - email
      - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
  responses:
    Unauthorized:
      description: Not logged in
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Role not allowed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /api/alumni/register:
    post:
      summary: Register and log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created; sets iw_token cookie
        '400':
          description: Validation failed
        '409':
          description: Email already registered
        '429':
          description: Rate limited
...
  /api/auth/login:
    post:
      summary: Log in (5 failures lock the account for 15 minutes)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK; sets iw_token cookie
        '400':
          description: Malformed input
        '401':
          description: Invalid credentials or account locked (deliberately identical)
        '429':
          description: Rate limited
...
  /api/auth/me:
    get:
      summary: Current user
      security:
      - cookieAuth: []
      responses:
        '200':
          description: OK
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/events/book:
    post:
      summary: Book an event (identity from session)
      security:
      - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - eventName
              - phone
              properties:
                eventName:
                  type: string
                phone:
                  type: string
      responses:
        '201':
          description: Booked
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Already booked
  /api/events/bookings:
    get:
      summary: All bookings (admin only)
      security:
      - cookieAuth: []
      responses:
        '200':
          description: OK
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/jobs:
    get:
      summary: 'List published jobs (public). Filters: q (text search), type, skill, location, page, limit<=20'
      responses:
        '200':
          description: OK
    post:
      summary: Post a job. Members' jobs are PENDING until an admin approves; employers/admins publish immediately
      responses:
        '200':
          description: OK
        '401': &id001
          $ref: '#/components/responses/Unauthorized'
        '201':
          description: Created
        '400':
          description: Validation failed
      security: &id002
      - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - title
              - company
              - location
              - type
              - description
              properties:
                title:
                  type: string
                company:
                  type: string
                location:
                  type: string
                type:
                  type: string
                  enum:
                  - full-time
                  - part-time
                  - contract
                  - internship
                  - remote
                description:
                  type: string
                  minLength: 30
                  maxLength: 5000
                skills:
                  type: array
                  items:
                    type: string
                  maxItems: 15
                applyUrl:
                  type: string
                  description: https only
...
  /api/applications/mine:
    get:
      summary: My applications and their status
      responses:
        '200':
          description: OK
        '401': *id001
      security: *id002
  /api/applications/{id}/status:
    patch:
      summary: 'Change application status (job owner/admin). Legal: submitted>reviewing|rejected, reviewing>shortlisted|rejected,
        shortlisted>hired|rejected'
      responses:
        '200':
          description: OK
        '401': *id001
        '404':
          description: Not found or not yours
        '409':
          description: Illegal transition
      security: *id002
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - status
              properties:
                status:
                  type: string
                  enum:
                  - reviewing
                  - shortlisted
                  - rejected
                  - hired
      parameters: *id003
  /api/admin/jobs/pending:
    get:
      summary: Moderation queue (admin)
      responses:
        '200':
          description: OK
        '401': *id001
        '403':
          $ref: '#/components/responses/Forbidden'
      security: *id002
  /api/admin/jobs/{id}/moderate:
    patch:
      summary: Approve or reject a pending job (admin)
      responses:
        '200':
          description: OK
        '401': *id001
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Already moderated
      security: *id002
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - action
              properties:
                action:
                  type: string
                  enum:
                  - approve
                  - reject
      parameters: *id003
  /api/alumni/me:
    get:
      summary: My matching profile (skills, occupation, bio)
      security: &id004
      - cookieAuth: []
      responses:
        '200':
          description: OK
        '401': &id005
          $ref: '#/components/responses/Unauthorized'
    patch:
      summary: Update my skills/occupation/bio (other fields ignored)
      security: *id004
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                skills:
                  type: array
                  items:
                    type: string
                  maxItems: 15
                occupation:
                  type: string
                  maxLength: 80
                bio:
                  type: string
                  maxLength: 500
      responses:
        '200':
          description: OK
        '400':
          description: Validation failed
        '401': *id005
...